started the blog, home, and login apps
The home app is where each post will be displayed. The database
will be polled for posts and displayed as they come in.
The login app will manage the creation, deletion, and site-addmittance
of users.
The blog app will handle creation, deletion, and editing of blogs
and will store them in the database.