Settup a homepage with bootstrap
I settup a view that will render a homepage when visiting the "/" url.
I included bootstrap as a static file for this homepage. It is likely
that I might need to make the static file usable by all apps, however
the home app will be doing the majority of the rendering so it may be
unnecessary to do this.