Added user login/logout, made bootstrap global
The user login/logout functionality works. There needs to be some
more error checking that is done and rendered on the logout.
I decided that I wanted some of the functionality of bootstrap
throughout the apps so I made bootstrap exist in the global static
directory along with a styles file that I wrote myself.